# backend/openapi.yaml
openapi: 3.0.3
info:
  title: HR / E-Learning API
  version: "1.0.0"
  description: |
    OpenAPI spec generated from your NestJS controllers & DTOs (User, Auth, Admin, Feedback, AuditLog, Backup).
    Sensitive fields (passwordHash, mfaSecret, mfaBackupCodes) are intentionally excluded from public response schemas.
servers:
  - url: http://localhost:3786
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          example: "Alice Example"
        email:
          type: string
          format: email
          example: "alice@example.com"
        password:
          type: string
          example: "P@ssw0rd!"
        role:
          type: string
          enum: [Student, Instructor, Admin]
          example: Student
        learningPreferences:
          type: array
          items:
            type: string
        subjectsOfInterest:
          type: array
          items:
            type: string
        expertise:
          type: array
          items:
            type: string
        profileImage:
          type: string
      required:
        - name
        - email
        - password
...
    LoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RegisterDto:
      $ref: '#/components/schemas/CreateUserDto'
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    AuthTokensDto:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOi..."
        refreshToken:
          type: string
          example: "def50200..."
    MfaSetupDto:
      type: object
      properties:
        otpauthUrl:
          type: string
        secret:
          type: string
    SimpleMessageDto:
      type: object
      properties:
        message:
          type: string
          example: "OTP sent to email"

    # -------------------------
    # Feedback
    # -------------------------
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered (may require email verification)
          content:
            application/json:
              schema:
                type: object
  /auth/verify-otp:
    post:
      tags:
        - auth
      summary: Verify registration/email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                otp:
                  type: string
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleMessageDto'
  /auth/login:
    post:
      tags:
        - auth
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensDto'
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Exchange refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensDto'
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout (invalidate current token)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleMessageDto'
  /auth/send-otp:
    post:
      tags:
        - auth
      summary: Send OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpDto'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleMessageDto'
  /auth/resend-otp:
    post:
      tags:
        - auth
      summary: Resend OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpDto'
      responses:
        '200':
          description: OTP resent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleMessageDto'
  /auth/otp-status/{email}:
    get:
      tags:
        - auth
      summary: Check OTP status for email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OTP status
          content:
            application/json:
              schema:
                type: object
  /auth/forgot-password:
    post:
      tags:
        - auth
      summary: Request forgot password OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleMessageDto'
  /auth/reset-password:
    post:
      tags:
        - auth
      summary: Reset password using OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleMessageDto'
  /auth/mfa/setup:
    post:
      tags:
        - auth
      summary: Initiate MFA setup (generate secret & otpauth)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: MFA setup info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaSetupDto'
  /auth/mfa/activate:
    post:
      tags:
        - auth
      summary: Activate MFA for current user (verify token)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaActivateDto'
      responses:
        '200':
          description: MFA activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "MFA enabled"
  /auth/mfa/verify-login:
    post:
      tags:
        - auth
      summary: Verify MFA during login (temp session)
      description: "Verify MFA during login flow. Provide the temporary token in Authorization header (bearer) as used by your TempJwtGuard."
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyLoginDto'
      responses:
        '200':
          description: Auth tokens after MFA verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensDto'
  /auth/mfa/disable:
    post:
      tags:
        - auth
      summary: Disable MFA for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: MFA disabled
          content:
            application/json:
              schema:
                type: object

  # -------------------------
  # User
  # -------------------------
...